+++ listenbrainz/messybrainz/README.md
@@ -1,0 +1,13 @@
+messybrainz-server
+==================
+
+The server components for the MessyBrainz project.
+
+MessyBrainz is a [MetaBrainz](https://metabrainz.org) project to support unclean metadata. While
+[MusicBrainz](https://musicbrainz.org) is designed to link clean metadata to stable identifiers,
+there is a need to identify unclean or misspelled data as well. MessyBrainz provides identifiers to
+unclean metadata, and where possible, links it to stable MusicBrainz identifiers.
+
+MessyBrainz is currently used in support of two projects, [ListenBrainz](https://listenbrainz.org)
+and [AcousticBrainz](https://acousticbrainz.org). Submission to MessyBrainz is restricted, however
+the resulting data will be made freely available.
